Renamed Internationalization js folder to i18n
