+++ assets/js/i18n/I18nLoader.js
