+++ assets/js/i18n/InternationalizedPage.js
