+++ assets/js/i18n/locales.js
