+++ assets/js/pages/lessons.page.js
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { createStore } from 'redux'
 import Navbar from '../Workflows/Navbar'
-import { localeReducer } from '../Internationalization/actions'
-import InternationalizedPage from '../Internationalization/InternationalizedPage'
+import { localeReducer } from '../i18n/actions'
+import InternationalizedPage from '../i18n/InternationalizedPage'
 
 const store = createStore(
   localeReducer
