+++ assets/js/pages/workflow.page.js
@@ -7,11 +7,11 @@ import UnhandledErrorReport from '../UnhandledErrorReport'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { workflowReducer, applyDeltaAction } from '../workflow-reducer'
-import { reducerFunctions as localeReducer } from '../Internationalization/actions'
+import { reducerFunctions as localeReducer } from '../i18n/actions'
 import Workflow from '../Workflow'
 import WorkflowWebsocket from '../WorkflowWebsocket'
 import WorkbenchAPI from '../WorkbenchAPI'
-import InternationalizedPage from '../Internationalization/InternationalizedPage'
+import InternationalizedPage from '../i18n/InternationalizedPage'
 
 //TODO: is this line needed? we have the info in webpack config, since it is nedded in dynamically fetching localization files
 __webpack_public_path__ = window.STATIC_URL + 'bundles/' // eslint-disable-line
