+++ assets/js/workflow-reducer.js
@@ -4,7 +4,7 @@ import { reducerFunctions as TabReducerFunctions } from './WorkflowEditor/Tabs/a
 import { reducerFunctions as WorkflowEditorReducerFunctions } from './WorkflowEditor/actions'
 import { reducerFunctions as ShareReducerFunctions } from './ShareModal/actions'
 import { reducerFunctions as FileReducerFunctions } from './params/File/actions'
-import { reducerFunctions as LocaleReducerFunctions } from './Internationalization/actions'
+import { reducerFunctions as LocaleReducerFunctions } from './i18n/actions'
 
 // Workflow
 const SET_WORKFLOW_NAME = 'SET_WORKFLOW_NAME'
